requirements.txt
streamlit==1.25.0
prophet==1.1.4
plotly==5.15.0 
yfinance==0.2.31